refactor: tighten cli parser maintenance boundaries
