+++ src/lifeos_cli/cli_support/resources/area/parser.py
@@ -5,7 +5,11 @@ from __future__ import annotations
 import argparse
 from uuid import UUID
 
-from lifeos_cli.cli_support.help_utils import HelpContent, add_documented_parser, make_help_handler
+from lifeos_cli.cli_support.help_utils import (
+    HelpContent,
+    add_documented_help_parser,
+    add_documented_parser,
+)
 from lifeos_cli.cli_support.output_utils import format_summary_column_list
 from lifeos_cli.cli_support.parser_common import (
     add_identifier_list_argument,
@@ -27,7 +31,7 @@ from lifeos_cli.i18n import gettext_message as _
 
 def build_area_parser(subparsers: argparse._SubParsersAction[argparse.ArgumentParser]) -> None:
     """Build the area command tree."""
-    area_parser = add_documented_parser(
+    area_parser = add_documented_help_parser(
         subparsers,
         "area",
         help_content=HelpContent(
@@ -55,7 +59,6 @@ def build_area_parser(subparsers: argparse._SubParsersAction[argparse.ArgumentPa
             ),
         ),
     )
-    area_parser.set_defaults(handler=make_help_handler(area_parser))
     area_subparsers = area_parser.add_subparsers(
         dest="area_command", title=_("actions"), metavar=_("action")
     )
@@ -213,7 +216,7 @@ def build_area_parser(subparsers: argparse._SubParsersAction[argparse.ArgumentPa
     delete_parser.add_argument("area_id", type=UUID, help=_("Area identifier"))
     delete_parser.set_defaults(handler=make_sync_handler(handle_area_delete_async))
 
-    batch_parser = add_documented_parser(
+    batch_parser = add_documented_help_parser(
         area_subparsers,
         "batch",
         help_content=HelpContent(
@@ -226,7 +229,6 @@ def build_area_parser(subparsers: argparse._SubParsersAction[argparse.ArgumentPa
             notes=(_("This namespace currently exposes only the `delete` workflow."),),
         ),
     )
-    batch_parser.set_defaults(handler=make_help_handler(batch_parser))
     batch_subparsers = batch_parser.add_subparsers(
         dest="area_batch_command",
         title=_("batch actions"),
