+++ src/lifeos_cli/cli_support/resources/habit_action/parser.py
@@ -5,7 +5,11 @@ from __future__ import annotations
 import argparse
 from uuid import UUID
 
-from lifeos_cli.cli_support.help_utils import HelpContent, add_documented_parser, make_help_handler
+from lifeos_cli.cli_support.help_utils import (
+    HelpContent,
+    add_documented_help_parser,
+    add_documented_parser,
+)
 from lifeos_cli.cli_support.output_utils import format_summary_column_list
 from lifeos_cli.cli_support.parser_common import (
     add_date_range_arguments,
@@ -28,7 +32,7 @@ def build_habit_action_parser(
     subparsers: argparse._SubParsersAction[argparse.ArgumentParser],
 ) -> None:
     """Build the habit-action command tree."""
-    action_parser = add_documented_parser(
+    action_parser = add_documented_help_parser(
         subparsers,
         "habit-action",
         help_content=HelpContent(
@@ -55,7 +59,6 @@ def build_habit_action_parser(
             ),
         ),
     )
-    action_parser.set_defaults(handler=make_help_handler(action_parser))
     action_subparsers = action_parser.add_subparsers(
         dest="habit_action_command",
         title=_("actions"),
