+++ src/lifeos_cli/cli_support/resources/note/parser.py
@@ -4,7 +4,10 @@ from __future__ import annotations
 
 import argparse
 
-from lifeos_cli.cli_support.help_utils import HelpContent, add_documented_parser, make_help_handler
+from lifeos_cli.cli_support.help_utils import (
+    HelpContent,
+    add_documented_help_parser,
+)
 from lifeos_cli.cli_support.resources.note.parser_actions import (
     build_note_add_parser,
     build_note_batch_parser,
@@ -19,7 +22,7 @@ from lifeos_cli.i18n import gettext_message as _
 
 def build_note_parser(subparsers: argparse._SubParsersAction[argparse.ArgumentParser]) -> None:
     """Build the note command tree."""
-    note_parser = add_documented_parser(
+    note_parser = add_documented_help_parser(
         subparsers,
         "note",
         help_content=HelpContent(
@@ -39,7 +42,6 @@ def build_note_parser(subparsers: argparse._SubParsersAction[argparse.ArgumentPa
             ),
         ),
     )
-    note_parser.set_defaults(handler=make_help_handler(note_parser))
     note_subparsers = note_parser.add_subparsers(
         dest="note_command",
         title=_("actions"),
