+++ src/lifeos_cli/cli_support/resources/people/parser.py
@@ -5,7 +5,11 @@ from __future__ import annotations
 import argparse
 from uuid import UUID
 
-from lifeos_cli.cli_support.help_utils import HelpContent, add_documented_parser, make_help_handler
+from lifeos_cli.cli_support.help_utils import (
+    HelpContent,
+    add_documented_help_parser,
+    add_documented_parser,
+)
 from lifeos_cli.cli_support.output_utils import format_summary_column_list
 from lifeos_cli.cli_support.parser_common import (
     add_identifier_list_argument,
@@ -27,7 +31,7 @@ from lifeos_cli.i18n import gettext_message as _
 
 def build_people_parser(subparsers: argparse._SubParsersAction[argparse.ArgumentParser]) -> None:
     """Build the people command tree."""
-    people_parser = add_documented_parser(
+    people_parser = add_documented_help_parser(
         subparsers,
         "people",
         help_content=HelpContent(
@@ -62,7 +66,6 @@ def build_people_parser(subparsers: argparse._SubParsersAction[argparse.Argument
             ),
         ),
     )
-    people_parser.set_defaults(handler=make_help_handler(people_parser))
     people_subparsers = people_parser.add_subparsers(
         dest="people_command", title=_("actions"), metavar=_("action")
     )
@@ -232,7 +235,7 @@ def build_people_parser(subparsers: argparse._SubParsersAction[argparse.Argument
     delete_parser.add_argument("person_id", type=UUID, help=_("Person identifier"))
     delete_parser.set_defaults(handler=make_sync_handler(handle_people_delete_async))
 
-    batch_parser = add_documented_parser(
+    batch_parser = add_documented_help_parser(
         people_subparsers,
         "batch",
         help_content=HelpContent(
@@ -245,7 +248,6 @@ def build_people_parser(subparsers: argparse._SubParsersAction[argparse.Argument
             notes=(_("This namespace currently exposes only the `delete` workflow."),),
         ),
     )
-    batch_parser.set_defaults(handler=make_help_handler(batch_parser))
     batch_subparsers = batch_parser.add_subparsers(
         dest="people_batch_command",
         title=_("batch actions"),
