+++ src/lifeos_cli/cli_support/system/config_commands.py
@@ -1,234 +1,25 @@
-"""Configuration and initialization CLI commands."""
+"""Configuration CLI command definitions."""
 
 from __future__ import annotations
 
 import argparse
-import sys
-from functools import partial
 
-from lifeos_cli.application.configuration import (
-    SUPPORTED_CONFIG_KEYS,
-    InitializationPrompts,
-    InitializationRequest,
-    build_database_settings,
-    build_preferences_settings,
-    persist_runtime_settings,
-    set_runtime_config_value,
-)
-from lifeos_cli.application.database import (
-    run_configured_database_subcommand_in_subprocess,
-)
-from lifeos_cli.cli_support import init_prompts
+from lifeos_cli.application.configuration import SUPPORTED_CONFIG_KEYS
 from lifeos_cli.cli_support.help_utils import (
     HelpContent,
+    add_documented_help_parser,
     add_documented_parser,
-    make_help_handler,
 )
-from lifeos_cli.cli_support.runtime_utils import (
-    format_config_summary,
-)
-from lifeos_cli.config import (
-    get_database_settings,
-    get_preferences_settings,
-    validate_database_schema_name,
-    validate_database_url,
-    validate_language,
+from lifeos_cli.cli_support.system.config_handlers import (
+    handle_config_set,
+    handle_config_show,
 )
 from lifeos_cli.i18n import gettext_message as _
 
 
-def _handle_init(args: argparse.Namespace) -> int:
-    """Initialize local configuration and verify database connectivity."""
-    request = InitializationRequest(
-        database_url=args.database_url,
-        schema=args.schema,
-        echo=args.echo,
-        timezone=args.timezone,
-        language=args.language,
-        day_starts_at=args.day_starts_at,
-        week_starts_on=args.week_starts_on,
-        vision_experience_rate_per_hour=args.vision_experience_rate_per_hour,
-        non_interactive=args.non_interactive,
-        is_interactive=sys.stdin.isatty(),
-        prompts=InitializationPrompts(
-            prompt_database_url=partial(
-                init_prompts.prompt_validated_text,
-                "Database URL",
-                validator=validate_database_url,
-            ),
-            prompt_database_schema=partial(
-                init_prompts.prompt_validated_text,
-                "Database schema",
-                validator=validate_database_schema_name,
-            ),
-            prompt_language=partial(
-                init_prompts.prompt_validated_text,
-                "Preferred language tag for human-authored payloads",
-                validator=validate_language,
-            ),
-            prompt_database_echo=partial(
-                init_prompts.prompt_bool,
-                _("Enable SQL echo logging"),
-            ),
-        ),
-    )
-    database_settings = build_database_settings(request)
-    preferences_settings = build_preferences_settings(request)
-    config_path = persist_runtime_settings(database_settings, preferences_settings)
-
-    print(f"Wrote config file: {config_path}")
-    print(
-        format_config_summary(
-            database_settings,
-            preferences_settings,
-            show_secrets=False,
-        )
-    )
-
-    if args.skip_ping:
-        print("Skipped database connectivity check.")
-    else:
-        run_configured_database_subcommand_in_subprocess(subcommand="ping")
-        print("Database connection succeeded.")
-
-    if args.skip_migrate:
-        print("Skipped database migrations. Run `lifeos db upgrade` when ready.")
-    else:
-        run_configured_database_subcommand_in_subprocess(subcommand="upgrade")
-        print("Database migrations are up to date.")
-
-    return 0
-
-
-def _handle_config_show(args: argparse.Namespace) -> int:
-    """Show the effective runtime configuration."""
-    print(
-        format_config_summary(
-            get_database_settings(),
-            get_preferences_settings(),
-            show_secrets=args.show_secrets,
-        )
-    )
-    return 0
-
-
-def _handle_config_set(args: argparse.Namespace) -> int:
-    """Persist one supported config key to the local config file."""
-    result = set_runtime_config_value(key=args.key, value=args.value)
-    print(f"Updated config file: {result.config_path}")
-    print(f"Updated key: {result.key}")
-    print(
-        format_config_summary(
-            result.database_settings,
-            result.preferences_settings,
-            show_secrets=args.show_secrets,
-        )
-    )
-    return 0
-
-
-def build_init_parser(subparsers: argparse._SubParsersAction[argparse.ArgumentParser]) -> None:
-    """Build the init command."""
-    init_parser = add_documented_parser(
-        subparsers,
-        "init",
-        help_content=HelpContent(
-            summary=_("Initialize local configuration"),
-            description=(
-                _(
-                    "Create or update the local LifeOS config file and verify that the "
-                    "database is reachable."
-                )
-                + "\n\n"
-                + _("This command is the recommended first step after installing lifeos-cli.")
-            ),
-            examples=(
-                "lifeos init",
-                "lifeos init --database-url "
-                "postgresql+psycopg://<db-user>:<db-password>@localhost:5432/lifeos",
-                "lifeos init --non-interactive --database-url "
-                "postgresql+psycopg://<db-user>:<db-password>@localhost:5432/lifeos "
-                "--timezone America/Toronto --language zh-Hans --skip-migrate",
-            ),
-            notes=(
-                _("Configuration is written to ~/.config/lifeos/config.toml by default."),
-                _("Environment variables still override config file values at runtime."),
-                _("Database credentials may be stored in plain text in the config file."),
-                _("Preference values are also stored in the config file under [preferences]."),
-                _(
-                    "Use `config set` for supported follow-up edits when you do not need to "
-                    "re-run the full bootstrap flow."
-                ),
-                _(
-                    "Interactive init explicitly confirms the language preference because agents "
-                    "should use it for human-authored payload data."
-                ),
-                _("Re-run `lifeos init` at any time to update stored preferences."),
-            ),
-        ),
-    )
-    init_parser.add_argument(
-        "--database-url",
-        help=_("PostgreSQL connection URL to persist in the config file"),
-    )
-    init_parser.add_argument(
-        "--schema",
-        default=None,
-        help=_("PostgreSQL schema name to use for application tables"),
-    )
-    init_parser.add_argument(
-        "--echo",
-        action=argparse.BooleanOptionalAction,
-        default=None,
-        help=_("Enable SQLAlchemy SQL echo logging in the config file"),
-    )
-    init_parser.add_argument(
-        "--timezone",
-        default=None,
-        help=_("Default IANA timezone for local day boundaries and time-based summaries"),
-    )
-    init_parser.add_argument(
-        "--language",
-        default=None,
-        help=_("Preferred language tag, for example en, en-CA, or zh-Hans"),
-    )
-    init_parser.add_argument(
-        "--day-starts-at",
-        default=None,
-        help=_("Local day boundary in HH:MM, used for future time-based grouping logic"),
-    )
-    init_parser.add_argument(
-        "--week-starts-on",
-        default=None,
-        help=_("Preferred first day of week: monday or sunday"),
-    )
-    init_parser.add_argument(
-        "--vision-experience-rate-per-hour",
-        type=int,
-        default=None,
-        help=_("Default vision experience points gained per hour of actual effort"),
-    )
-    init_parser.add_argument(
-        "--non-interactive",
-        action="store_true",
-        help=_("Do not prompt for missing values; require flags or existing config values"),
-    )
-    init_parser.add_argument(
-        "--skip-ping",
-        action="store_true",
-        help=_("Do not check database connectivity after writing the config file"),
-    )
-    init_parser.add_argument(
-        "--skip-migrate",
-        action="store_true",
-        help=_("Do not run database migrations after writing the config file"),
-    )
-    init_parser.set_defaults(handler=_handle_init)
-
-
 def build_config_parser(subparsers: argparse._SubParsersAction[argparse.ArgumentParser]) -> None:
     """Build the config command tree."""
-    config_parser = add_documented_parser(
+    config_parser = add_documented_help_parser(
         subparsers,
         "config",
         help_content=HelpContent(
@@ -256,7 +47,6 @@ def build_config_parser(subparsers: argparse._SubParsersAction[argparse.Argument
             ),
         ),
     )
-    config_parser.set_defaults(handler=make_help_handler(config_parser))
     config_subparsers = config_parser.add_subparsers(
         dest="config_command",
         title=_("actions"),
@@ -295,7 +85,7 @@ def build_config_parser(subparsers: argparse._SubParsersAction[argparse.Argument
         action="store_true",
         help=_("Print sensitive values such as database passwords in full"),
     )
-    show_parser.set_defaults(handler=_handle_config_show)
+    show_parser.set_defaults(handler=handle_config_show)
 
     set_parser = add_documented_parser(
         config_subparsers,
@@ -330,4 +120,4 @@ def build_config_parser(subparsers: argparse._SubParsersAction[argparse.Argument
         action="store_true",
         help=_("Print sensitive values such as database passwords in full after writing"),
     )
-    set_parser.set_defaults(handler=_handle_config_set)
+    set_parser.set_defaults(handler=handle_config_set)
